requirements.txt
django
djangorestframework
django-rest-knox
mysqlclient
django-cors-headers
mysqlclient
